Correct includes value in library.properties
The includes field in library.properties is used to specify which #include directives should be added to the sketch via Sketch > Include Library > IotWebConf. This field should specify the primary header file(s) of the library.

The includes field is only used by Arduino IDE 1.6.10 and newer and these IDE versions do not require #include directives in the sketch for resolution of dependencies of libraries so having ESP8266WiFi.h,ESP8266WebServer.h,DNSServer.h in the includes field is not necessary.