Make loadConfig() public
init() does several other things, such as setting up pins and mDNS
(which in turn starts the Webserver). This is not desired for some
applications which just need the settings and want to run as little
code as possible (e.g. because they don't need Wifi all the time and run
on battery).