Fixed broken uncheck of checkboxes
Added a method to indicate if a parameter shall be
updated even if parameter is missing in the web POST request.
WHen unchecking a checkbox, nothing is sent for
that checkbox in the web request and thus it was
not possible to save when a checkbox had been unchecked,
resulting in checkboxes always being checked.